+++ src/models/base/Oauth2AuthCodeScope.php
@@ -54,14 +54,14 @@ abstract class Oauth2AuthCodeScope extends \rhertogh\Yii2Oauth2Server\models\bas
      * @return \rhertogh\Yii2Oauth2Server\interfaces\models\queries\Oauth2AuthCodeQueryInterface     */
     public function getAuthCode()
     {
-        return $this->hasOne(\rhertogh\Yii2Oauth2Server\models\Oauth2AuthCode::className(), ['id' => 'auth_code_id'])->inverseOf('authCodeScopes');
+        return $this->hasOne(\rhertogh\Yii2Oauth2Server\models\Oauth2AuthCode::class, ['id' => 'auth_code_id'])->inverseOf('authCodeScopes');
     }
     
     /**
      * @return \rhertogh\Yii2Oauth2Server\interfaces\models\queries\Oauth2ScopeQueryInterface     */
     public function getScope()
     {
-        return $this->hasOne(\rhertogh\Yii2Oauth2Server\models\Oauth2Scope::className(), ['id' => 'scope_id'])->inverseOf('authCodeScopes');
+        return $this->hasOne(\rhertogh\Yii2Oauth2Server\models\Oauth2Scope::class, ['id' => 'scope_id'])->inverseOf('authCodeScopes');
     }