+++ src/models/base/Oauth2RefreshToken.php
@@ -64,7 +64,7 @@ abstract class Oauth2RefreshToken extends \rhertogh\Yii2Oauth2Server\models\base
      * @return \rhertogh\Yii2Oauth2Server\interfaces\models\queries\Oauth2AccessTokenQueryInterface     */
     public function getAccessToken()
     {
-        return $this->hasOne(\rhertogh\Yii2Oauth2Server\models\Oauth2AccessToken::className(), ['id' => 'access_token_id'])->inverseOf('refreshTokens');
+        return $this->hasOne(\rhertogh\Yii2Oauth2Server\models\Oauth2AccessToken::class, ['id' => 'access_token_id'])->inverseOf('refreshTokens');
     }